Conversion to Primarily 64INT, Optimizations

+++ _co_consts.c
@@ -1,7 +1,7 @@
 #include "assembler.h"
 
-int16_t search_const_table(Value val) {
-    for (int32_t i = 0; i < _co_consts_size; i++) {
+uint64_t search_const_table(Value val) {
+    for (uint64_t i = 0; i < _co_consts_size; i++) {
         if (val.type == INT32 && _co_consts[i].value.i32 == val.value.i32) { return i; }
         else if (val.type == INT64 && _co_consts[i].value.i64 == val.value.i64) { return i; }
         else if (val.type == CHAR && _co_consts[i].value.chr == val.value.chr) { return i; }
