+++ _code.c
@@ -5,7 +5,7 @@ void add_op_code(Function* func, Opcodes val) {
     (*func).code[(*func).code_size++] = (uint8_t) val;
 }
 
-void add_oper_code(Function* func, int8_t val) {
+void add_oper_code(Function* func, int64_t val) {
     (*func).code = realloc((*func).code, ((*func).code_size)+OPERAND_SIZE);
     memcpy(&(*func).code[(*func).code_size], &val, OPERAND_SIZE);
     (*func).code_size += OPERAND_SIZE;
